+++ R/BEDMatrix.R
@@ -1,60 +1,7 @@
-# Delimiters used in PED files.
+# Delimiters used in PED files
 delims <- "[ \t]"
 
-#' Creates a matrix wrapper around binary PED files.
-#'
-#' \code{BEDMatrix} is an S3 class that behaves similarly to a regular
-#' \code{matrix} by implementing key methods such as \code{[}, \code{dim}, and
-#' \code{dimnames}. Subsets are extracted directly and on-demand from the
-#' binary PED file without loading the entire file into memory through memory
-#' mapping. The subsets are coded similarly to RAW files generated with the
-#' \code{--recodeA} argument in PLINK: \code{0} indicates homozygous major
-#' allele, \code{1} indicates heterozygous, and \code{2} indicates homozygous
-#' minor allele.
-#'
-#' A \code{BEDMatrix} instance can be created by providing the path to the BED
-#' file (with or without extension) as \code{path}, the number of individuals
-#' as \code{n}, and the number of markers as \code{p}. If a FAM file (which
-#' corresponds to the first six columns of a PED file) of the same name and in
-#' the same directory as the BED file exists, it is optional to provide
-#' \code{n} and the number of individuals as well as the rownames of the
-#' \code{BEDMatrix} will be detected automatically. The rownames will be
-#' generated based on the IID and FID of each individual, concatenated by
-#' \code{_}.  If a BIM file (which corresponds to the MAP file that accompanies
-#' a PED file) of the same name and in the same directory as the BED file
-#' exists, it is optional to provide \code{p} and the number of markers as well
-#' as the colnames of the \code{BEDMatrix} will be detected automatically. The
-#' colnames will be generated based on the SNP name and the minor allele,
-#' concatenated by \code{_} (similar to the colnames in a RAW file). For very
-#' large BED file it is advised to provide \code{n} and \code{p} manually to
-#' speed up object creation. In that case \code{rownames} and \code{colnames}
-#' will be set to \code{NULL} and have to be specified manually.
-#'
-#' A BED file can be created from a PED file with
-#' \href{http://pngu.mgh.harvard.edu/~purcell/plink/}{PLINK} using \code{plink
-#' --file myfile --make-bed}. BED files are storage and query efficient, and can
-#' be transformed back into the original PED file with PLINK using \code{plink
-#' --bfile myfile --recode}.
-#'
-#' Internally, \code{BEDMatrix} inherits from \code{list} and
-#' exposes a few attributes that should not be relied upon in actual code:
-#' \code{path}, \code{dims}, \code{dnames}, and \code{_instance}. \code{path}
-#' stores the path to the BED file. \code{dims} and \code{dnames} contain the
-#' dimensions and dimnames of the BEDMatrix object. \code{_instance} points to
-#' the underlying \code{Rcpp} module. The \code{Rcpp} module exposes an S4 class
-#' called \code{BEDMatrix_} that memory maps the BED file via
-#' \code{Boost.Interprocess} of the \code{BH} package.
-#'
-#' @param path Path to the binary PED file, with or without extension.
-#' @param n The number of individuals. Optional if FAM file of same name as BED
-#'   file exists. If provided, \code{rownames} will be set to \code{NULL} and
-#'   have to be provided manually.
-#' @param p The number of markers. Optional if BIM file of same name as BED file
-#'   exists. If provided, \code{colnames} will be set to \code{NULL} and have
-#'   to be provided manually.
-#' @export
-#' @example man/examples/BEDMatrix.R
-BEDMatrix <- function(path, n = NULL, p = NULL) {
+initialize <- function(.Object, path, n = NULL, p = NULL) {
     path <- path.expand(path)
     if (!file.exists(path)) {
         # Try to add extension (common in PLINK)
@@ -63,20 +10,20 @@ BEDMatrix <- function(path, n = NULL, p = NULL) {
             stop("File not found.")
         }
     }
-    dir <- substr(path, 1, nchar(path) - 4)
+    dir <- substr(path, 1L, nchar(path) - 4L)
     if (is.null(n)) {
         # Check if FAM file exists
-        if (!file.exists(paste0(dir, ".fam"))) {
-            stop("FAM file of same name not found. Provide number of individuals (n).")
+        famPath <- paste0(dir, ".fam")
+        if (!file.exists(famPath)) {
+            stop("FAM file of same name not found. Provide number of samples (n).")
         } else {
-            message("Extracting number of individuals and rownames from FAM file...")
-            famPath <- paste0(dir, ".fam")
+            message("Extracting number of samples and rownames from FAM file...")
             if (requireNamespace("data.table", quietly = TRUE)) {
-                fam <- data.table::fread(famPath, select = c(1, 2), data.table = FALSE, showProgress = FALSE)
+                fam <- data.table::fread(famPath, select = c(1L, 2L), data.table = FALSE, showProgress = FALSE)
                 # Determine n
                 n <- nrow(fam)
                 # Determine rownames
-                rownames <- paste0(fam[, 1], "_", fam[, 2])
+                rownames <- paste0(fam[, 1L], "_", fam[, 2L])
             } else {
                 fam <- readLines(famPath)
                 # Determine n
@@ -84,7 +31,7 @@ BEDMatrix <- function(path, n = NULL, p = NULL) {
                 # Determine rownames
                 rownames <- sapply(strsplit(fam, delims), function(line) {
                     # Concatenate family ID and subject ID
-                    return(paste0(line[1], "_", line[2]))
+                    return(paste0(line[1L], "_", line[2L]))
                 })
             }
         }
@@ -94,17 +41,17 @@ BEDMatrix <- function(path, n = NULL, p = NULL) {
     }
     if (is.null(p)) {
         # Check if BIM file exists
-        if (!file.exists(paste0(dir, ".bim"))) {
-            stop("BIM file of same name not found. Provide number of markers (p).")
+        bimPath <- paste0(dir, ".bim")
+        if (!file.exists(bimPath)) {
+            stop("BIM file of same name not found. Provide number of variants (p).")
         } else {
-            message("Extracting number of markers and colnames from BIM file...")
-            bimPath <- paste0(dir, ".bim")
+            message("Extracting number of variants and colnames from BIM file...")
             if (requireNamespace("data.table", quietly = TRUE)) {
-                bim <- data.table::fread(bimPath, select = c(2, 5), data.table = FALSE, showProgress = FALSE)
+                bim <- data.table::fread(bimPath, select = c(2L, 5L), data.table = FALSE, showProgress = FALSE)
                 # Determine p
                 p <- nrow(bim)
                 # Determine colnames
-                colnames <- paste0(bim[, 1], "_", bim[, 2])
+                colnames <- paste0(bim[, 1L], "_", bim[, 2L])
             } else {
                 bim <- readLines(bimPath)
                 # Determine p
@@ -112,7 +59,7 @@ BEDMatrix <- function(path, n = NULL, p = NULL) {
                 # Determine colnames
                 colnames <- sapply(strsplit(bim, delims), function(line) {
                     # Concatenate SNP name and minor allele (like --recodeA)
-                    return(paste0(line[2], "_", line[5]))
+                    return(paste0(line[2L], "_", line[5L]))
                 })
             }
         }
@@ -121,102 +68,169 @@ BEDMatrix <- function(path, n = NULL, p = NULL) {
         colnames <- NULL
     }
     # Create Rcpp object
-    rcpp_obj <- new(BEDMatrix_, path, n, p)
-    # Wrap object in S3 class
-    s3_obj <- list()
-    class(s3_obj) <- "BEDMatrix"
-    attr(s3_obj, "_instance") <- rcpp_obj
-    attr(s3_obj, "dnames") <- list(rownames, colnames)
-    attr(s3_obj, "dims") <- c(rcpp_obj$n, rcpp_obj$p)
-    attr(s3_obj, "path") <- path
-    return(s3_obj)
+    .Object@xptr <- .Call("BEDMatrix__new", path, n, p)
+    .Object@path <- path
+    .Object@dims <- c(n, p)
+    .Object@dnames <- list(rownames, colnames)
+    return(.Object)
 }
 
-#' @export
-print.BEDMatrix <- function(x, ...) {
-    dims <- dim(x)
-    n <- dims[1]
-    p <- dims[2]
-    cat("BEDMatrix: ", n, " x ", p, " [", attr(x, "path"), "]\n", sep = "")
+extract_vector <- function(x, i) {
+    .Call("BEDMatrix__extract_vector", x@xptr, i)
 }
 
-#' @export
-str.BEDMatrix <- function(object, ...) {
-    print(object)
-}
-
-#' @export
-`[.BEDMatrix` <- function(x, i, j, drop = TRUE) {
-    rcpp_obj <- attr(x, "_instance")
-    dims <- dim(x)
-    n <- dims[1]
-    p <- dims[2]
-    if (nargs() > 2) {
-        # Case [i, j]
-        if (missing(i)) {
-            i <- 1:n
-        } else if (class(i) == "logical") {
-            i <- which(rep_len(i, n))
-        } else if (class(i) == "character") {
-            i <- match(i, rownames(x))
-        }
-        if (missing(j)) {
-            j <- 1:p
-        } else if (class(j) == "logical") {
-            j <- which(rep_len(j, p))
-        } else if (class(j) == "character") {
-            j <- match(j, colnames(x))
-        }
-        subset <- rcpp_obj$matrixSubset(x, i, j)
-        # Let R handle drop behavior
-        if (drop == TRUE && (nrow(subset) == 1 || ncol(subset) == 1)) {
-            subset <- subset[, ]
-        }
-    } else {
-        if (missing(i)) {
-            # Case []
-            i <- 1:n
-            j <- 1:p
-            subset <- rcpp_obj$matrixSubset(x, i, j)
-        } else {
-            # Case [i]
-            if (class(i) == "matrix") {
-                i <- as.vector(i)
-                if (class(i) == "logical") {
-                  i <- which(rep_len(i, n * p))
-                  # matrix treats NAs as TRUE
-                  i <- sort(c(i, which(is.na(x[]))))
-                }
-            } else {
-                if (class(i) == "logical") {
-                  i <- which(rep_len(i, n * p))
-                }
-            }
-            subset <- rcpp_obj$vectorSubset(x, i)
-        }
-    }
+extract_matrix <- function(x, i, j) {
+    subset <- .Call("BEDMatrix__extract_matrix", x@xptr, i, j)
+    # Preserve dimnames
+    names <- x@dnames
+    dimnames(subset) <- list(
+        names[[1L]][i],
+        names[[2L]][j]
+    )
     return(subset)
 }
 
+show <- function(object) {
+    dims <- dim(object)
+    n <- dims[1L]
+    p <- dims[2L]
+    cat("BEDMatrix: ", n, " x ", p, " [", object@path, "]\n", sep = "")
+}
+
+#' A Class to Extract Genotypes from a PLINK .bed File.
+#'
+#' `BEDMatrix` is a class that maps a [PLINK .bed](https://www.cog-genomics.org/plink2/formats#bed)
+#' file into memory and behaves similarly to a regular `matrix` by implementing
+#' key methods such as `[`, `dim`, and `dimnames`. Subsets are extracted
+#' directly and on-demand from the .bed file without loading the entire file
+#' into memory.
+#'
+#' The subsets extracted from a `BEDMatrix` object are coded similarly to
+#' [.raw](https://www.cog-genomics.org/plink2/formats#raw) files (generated
+#' with the `--recodeA` argument in
+#' [PLINK](https://www.cog-genomics.org/plink2/)): `0` indicates homozygous
+#' major allele, `1` indicates heterozygous, and `2` indicates homozygous minor
+#' allele.
+#'
+#' Internally, this class is an S4 class with the following slots that should
+#' not be relied upon in actual code: `xptr`, `dims`, `dnames`, and `path`. The
+#' .bed file is mapped into memory using the [Rcpp][Rcpp::Rcpp-package] package
+#' and the `Boost.Interprocess` library provided by the [BH][BH::BH-package]
+#' package.
+#'
+#' @section Methods:
+#' - `[`
+#' - `dim`
+#' - `dimnames`
+#' - `dimnames<-`
+#' - `as.matrix`
+#' - `is.matrix`
+#' - `length`
+#' - `str`
+#' - `show`
+#' - `initialize`
+#'
+#' @slot xptr An external pointer to the underlying [Rcpp][Rcpp::Rcpp-package]
+#' code.
+#' @slot dims An integer vector specifying the number of samples and variants
+#' as determined by the the accompanying
+#' [.fam](https://www.cog-genomics.org/plink2/formats#fam) and
+#' [.bim](https://www.cog-genomics.org/plink2/formats#bim) files or by the `n`
+#' and `p` parameters of the [constructor
+#' function][initialize,BEDMatrix-method()].
+#' @slot dnames A list describing the row names and column names of the object
+#' as determined by the accompanying
+#' [.fam](https://www.cog-genomics.org/plink2/formats#fam) and
+#' [.bim](https://www.cog-genomics.org/plink2/formats#bim) files, or `NULL` if
+#' the `n` and `p` parameters of the [constructor
+#' function][initialize,BEDMatrix-method()] were provided.
+#' @slot path A character string containing the path to the .bed file.
+#' @seealso [initialize()][initialize,BEDMatrix-method()] to create a
+#' `BEDMatrix` object from a .bed file, [BEDMatrix-package] to learn more about
+#' .bed files, [LinkedMatrix][LinkedMatrix::LinkedMatrix-package] to link
+#' several `BEDMatrix` objects together.
+#' @example man/examples/BEDMatrix.R
+#' @aliases BEDMatrix-class
+#' @export BEDMatrix
+#' @exportClass BEDMatrix
+BEDMatrix <- setClass("BEDMatrix", slots = c(xptr = "externalptr", dims = "integer", dnames = "list", path = "character"))
+
+#' Create a BEDMatrix Object from a PLINK .bed File.
+#'
+#' This function constructs a new [BEDMatrix-class] object by mapping the
+#' specified [PLINK .bed](https://www.cog-genomics.org/plink2/formats#bed) file
+#' into memory.
+#'
+#' [.bed](https://www.cog-genomics.org/plink2/formats#bed) files must be
+#' accompanied by [.fam](https://www.cog-genomics.org/plink2/formats#fam) and
+#' [.bim](https://www.cog-genomics.org/plink2/formats#bim) files: .fam files
+#' contain sample information, and .bim files contain variant information. If
+#' the name of the .bed file is *plink*.bed then the names of the .fam and .bim
+#' files have to be *plink*.fam and *plink*.bim, respectively. The .fam and
+#' .bim files are used to extract the number and names of samples and variants.
+#'
+#' For very large .bed files, reading the .fam and .bim files can take a long
+#' time. If `n` and `p` are provided, these files are not read and `dimnames`
+#' have to be provided manually.
+#'
+#' Currently, only the variant-major mode of .bed files is supported.
+#' [PLINK2](https://www.cog-genomics.org/plink2/) "dropped" support for the
+#' sample-major mode by automatically converting files in this format to the
+#' variant-major mode. Therefore, it is recommended to run files in
+#' sample-major mode through PLINK2 first.
+#'
+#' @param .Object Internal, used by [methods::initialize()] generic.
+#' @param path Path to the
+#' [.bed](https://www.cog-genomics.org/plink2/formats#bed) file (with or
+#' without extension).
+#' @param n The number of samples. If `NULL` (the default), this number will be
+#' determined from the accompanying
+#' [.fam](https://www.cog-genomics.org/plink2/formats#fam) file (of the same
+#' name as the [.bed](https://www.cog-genomics.org/plink2/formats#bed) file).
+#' If a positive integer, the .fam file is not read and `rownames` will be set
+#' to `NULL` and have to be provided manually.
+#' @param p The number of variants. If `NULL` (the default) the number of
+#' variants will be determined from the accompanying
+#' [.bim](https://www.cog-genomics.org/plink2/formats#bim) file (of the same
+#' name as the [.bed](https://www.cog-genomics.org/plink2/formats#bed) file).
+#' If a positive integer, the .bim file is not read and `colnames` will be set
+#' to `NULL` and have to be provided manually.
+#' @return A [BEDMatrix-class] object.
+#' @example man/examples/initialize.R
+#' @seealso [BEDMatrix-package] to learn more about .bed files.
+#' @export
+setMethod("initialize", signature(.Object = "BEDMatrix"), initialize)
+
+#' Show a BEDMatrix Object.
+#'
+#' Display the object, by printing, plotting or whatever suits its class.
+#'
+#' @param object A [BEDMatrix-class] object.
+#' @export
+setMethod("show", signature(object = "BEDMatrix"), show)
+
+#' @export
+`[.BEDMatrix` <- crochet::extract(extract_vector = extract_vector, extract_matrix = extract_matrix)
+
 #' @export
 dim.BEDMatrix <- function(x) {
-    attr(x, "dims")
+    x@dims
 }
 
 #' @export
 dimnames.BEDMatrix <- function(x) {
-    attr(x, "dnames")
+    x@dnames
 }
 
 #' @export
 `dimnames<-.BEDMatrix` <- function(x, value) {
     d <- dim(x)
-    v1 <- value[[1]]
-    v2 <- value[[2]]
-    if (!is.list(value) || length(value) != 2 || !(is.null(v1) || length(v1) == d[1]) || !(is.null(v2) || length(v2) == d[2])) {
+    v1 <- value[[1L]]
+    v2 <- value[[2L]]
+    if (!is.list(value) || length(value) != 2L || !(is.null(v1) || length(v1) == d[1L]) || !(is.null(v2) || length(v2) == d[2L])) {
         stop("invalid dimnames")
     }
-    attr(x, "dnames") <- lapply(value, function(v) {
+    x@dnames <- lapply(value, function(v) {
         if (!is.null(v)) {
             as.character(v)
         }
@@ -230,11 +244,16 @@ length.BEDMatrix <- function(x) {
 }
 
 #' @export
-is.matrix.BEDMatrix <- function(x) {
-    TRUE
+str.BEDMatrix <- function(object, ...) {
+    print(object)
 }
 
 #' @export
 as.matrix.BEDMatrix <- function(x, ...) {
     x[, , drop = FALSE]
+}
+
+#' @export
+is.matrix.BEDMatrix <- function(x) {
+    TRUE
 }