+++ src/BEDMatrix.cpp
@@ -9,31 +9,11 @@
 #include <Rcpp.h>
 #include <string>
 
-Rcpp::IntegerMatrix& preserve_dimnames(const Rcpp::List& x, Rcpp::IntegerMatrix& out, const Rcpp::IntegerVector& i, const Rcpp::IntegerVector& j) {
-    Rcpp::List out_dimnames = Rcpp::List::create(
-        R_NilValue,
-        R_NilValue
-    );
-    Rcpp::List in_dimnames = x.attr("dnames");
-    Rcpp::RObject in_rownames = in_dimnames[0];
-    Rcpp::RObject in_colnames = in_dimnames[1];
-    if (!in_rownames.isNULL()) {
-        out_dimnames[0] = Rcpp::CharacterVector(in_rownames)[i];
-    }
-    if (!in_colnames.isNULL()) {
-        out_dimnames[1] = Rcpp::CharacterVector(in_colnames)[j];
-    }
-    out.attr("dimnames") = out_dimnames;
-    return out;
-}
-
 class BEDMatrix {
     public:
         BEDMatrix(std::string path, std::size_t n, std::size_t p);
-        Rcpp::IntegerVector vector_subset(Rcpp::List x, Rcpp::IntegerVector i);
-        Rcpp::IntegerMatrix matrix_subset(Rcpp::List x, Rcpp::IntegerVector i, Rcpp::IntegerVector j);
-        std::size_t get_nrow();
-        std::size_t get_ncol();
+        Rcpp::IntegerVector extract_vector(Rcpp::IntegerVector i);
+        Rcpp::IntegerMatrix extract_matrix(Rcpp::IntegerVector i, Rcpp::IntegerVector j);
     private:
         BEDMatrix(const BEDMatrix&);
         BEDMatrix& operator=(const BEDMatrix&);
@@ -51,27 +31,27 @@ BEDMatrix::BEDMatrix(std::string path, std::size_t n, std::size_t p) : nrow(n), 
     try {
         this->file = boost::interprocess::file_mapping(path.c_str(), boost::interprocess::read_only);
     } catch(const boost::interprocess::interprocess_exception& e) {
-        Rcpp::stop("File not found.");
+        throw std::runtime_error("File not found.");
     }
     this->file_region = boost::interprocess::mapped_region(this->file, boost::interprocess::read_only);
     this->file_data = static_cast<const char*>(this->file_region.get_address());
     // Check magic number
     if (!(this->file_data[0] == '\x6C' && this->file_data[1] == '\x1B')) {
-        Rcpp::stop("File is not a binary PED file.");
+        throw std::runtime_error("File is not a binary PED file.");
     }
-    // Check mode: 00000001 indicates the default SNP-major mode (i.e.
-    // list all individuals for first SNP, all individuals for second
-    // SNP, etc), 00000000 indicates the unsupported individual-major
-    // mode (i.e. list all SNPs for the first individual, list all SNPs
-    // for the second individual, etc)
+    // Check mode: 00000001 indicates the default variant-major mode (i.e.
+    // list all samples for first variant, all samples for second variant,
+    // etc), 00000000 indicates the unsupported sample-major mode (i.e. list
+    // all variants for the first sample, list all variants for the second
+    // sample, etc)
     if (this->file_data[2] != '\x01') {
-        Rcpp::stop("Individual-major mode is not supported.");
+        throw std::runtime_error("Sample-major mode is not supported.");
     }
     // Get number of bytes
     const std::size_t num_bytes = this->file_region.get_size();
     // Check if given dimensions match the file
     if ((this->nrow * this->ncol) + (this->byte_padding * this->ncol) != (num_bytes - this->length_header) * 4) {
-        Rcpp::stop("n or p does not match the dimensions of the file.");
+        throw std::runtime_error("n or p does not match the dimensions of the file.");
     }
 }
 
@@ -102,28 +82,30 @@ int BEDMatrix::get_genotype(std::size_t i, std::size_t j) {
     return mapping;
 }
 
-Rcpp::IntegerVector BEDMatrix::vector_subset(Rcpp::List x, Rcpp::IntegerVector i) {
-    // Check if index is out of bounds
-    if (Rcpp::is_true(Rcpp::any(i > this->nrow * this->ncol))) {
-        Rcpp::stop("Invalid dimensions.");
-    }
+Rcpp::IntegerVector BEDMatrix::extract_vector(Rcpp::IntegerVector i) {
     // Convert from 1-index to 0-index
     Rcpp::IntegerVector i0(i - 1);
     // Keep size of i
     std::size_t size_i = i.size();
     // Reserve output vector
     Rcpp::IntegerVector out(size_i);
+    // Get bounds
+    std::size_t bounds = this->nrow * this->ncol;
     // Iterate over indexes
     for (std::size_t idx_i = 0; idx_i < size_i; idx_i++) {
-        out(idx_i) = this->get_genotype(i0[idx_i] % this->nrow, i0[idx_i] / this->nrow);
+        if (Rcpp::IntegerVector::is_na(i0[idx_i]) || static_cast<std::size_t>(i0[idx_i]) >= bounds) {
+            out(idx_i) = NA_INTEGER;
+        } else {
+            out(idx_i) = this->get_genotype(i0[idx_i] % this->nrow, i0[idx_i] / this->nrow);
+        }
     }
     return out;
 }
 
-Rcpp::IntegerMatrix BEDMatrix::matrix_subset(Rcpp::List x, Rcpp::IntegerVector i, Rcpp::IntegerVector j) {
+Rcpp::IntegerMatrix BEDMatrix::extract_matrix(Rcpp::IntegerVector i, Rcpp::IntegerVector j) {
     // Check if indexes are out of bounds
     if (Rcpp::is_true(Rcpp::any(i > this->nrow)) || Rcpp::is_true(Rcpp::any(j > this->ncol))) {
-        Rcpp::stop("Invalid dimensions.");
+        throw std::runtime_error("subscript out of bounds");
     }
     // Convert from 1-index to 0-index
     Rcpp::IntegerVector i0(i - 1);
@@ -133,38 +115,67 @@ Rcpp::IntegerMatrix BEDMatrix::matrix_subset(Rcpp::List x, Rcpp::IntegerVector i
     std::size_t size_j = j.size();
     // Reserve output matrix
     Rcpp::IntegerMatrix out(size_i, size_j);
-    preserve_dimnames(x, out, i0, j0);
     // Iterate over column indexes
     for (std::size_t idx_j = 0; idx_j < size_j; idx_j++) {
         // Iterate over row indexes
         for (std::size_t idx_i = 0; idx_i < size_i; idx_i++) {
-            out(idx_i, idx_j) = this->get_genotype(i0[idx_i], j0[idx_j]);
+            if (Rcpp::IntegerVector::is_na(i0[idx_i]) || Rcpp::IntegerVector::is_na(j0[idx_j])) {
+                out(idx_i, idx_j) = NA_INTEGER;
+            } else {
+                out(idx_i, idx_j) = this->get_genotype(i0[idx_i], j0[idx_j]);
+            }
         }
     }
     return out;
 }
 
-
-std::size_t BEDMatrix::get_nrow() {
-    return this->nrow;
-};
-
-std::size_t BEDMatrix::get_ncol() {
-    return this->ncol;
-};
-
 const unsigned short int BEDMatrix::length_header = 3;
 
-RCPP_MODULE(mod_BEDMatrix) {
+// Export BEDMatrix::BEDMatrix
+RcppExport SEXP BEDMatrix__new(SEXP path_, SEXP n_, SEXP p_) {
+    // Convert inputs to appropriate C++ types
+    std::string path = Rcpp::as<std::string>(path_);
+    std::size_t n = Rcpp::as<std::size_t>(n_);
+    std::size_t p = Rcpp::as<std::size_t>(p_);
+    try {
+        // Create a pointer to a BEDMatrix object and wrap it as an external
+        // pointer
+        Rcpp::XPtr<BEDMatrix> ptr(new BEDMatrix(path, n, p), true);
+        // Return the external pointer to the R side
+        return ptr;
+    } catch(std::exception &ex) {
+        forward_exception_to_r(ex);
+        return 0;
+    }
+};
 
-    using namespace Rcpp;
+// Export BEDMatrix::extract_vector
+RcppExport SEXP BEDMatrix__extract_vector(SEXP xp_, SEXP i_) {
+    // Convert inputs to appropriate C++ types
+    Rcpp::XPtr<BEDMatrix> ptr(xp_);
+    Rcpp::IntegerVector i = Rcpp::as<Rcpp::IntegerVector>(i_);
+    try {
+        // Invoke the extract_vector function
+        Rcpp::IntegerVector res = ptr->extract_vector(i);
+        return res;
+    } catch(std::exception &ex) {
+        forward_exception_to_r(ex);
+        return 0;
+    }
+};
 
-    class_<BEDMatrix>("BEDMatrix_")
-    .constructor<std::string, std::size_t, std::size_t>()
-    .method("vectorSubset", &BEDMatrix::vector_subset)
-    .method("matrixSubset", &BEDMatrix::matrix_subset)
-    .property("n", &BEDMatrix::get_nrow)
-    .property("p", &BEDMatrix::get_ncol)
-    ;
-
-}
+// Export BEDMatrix::extract_matrix
+RcppExport SEXP BEDMatrix__extract_matrix(SEXP xp_, SEXP i_, SEXP j_) {
+    // Convert inputs to appropriate C++ types
+    Rcpp::XPtr<BEDMatrix> ptr(xp_);
+    Rcpp::IntegerVector i = Rcpp::as<Rcpp::IntegerVector>(i_);
+    Rcpp::IntegerVector j = Rcpp::as<Rcpp::IntegerVector>(j_);
+    try {
+        // Invoke the extract_matrix function
+        Rcpp::IntegerMatrix res = ptr->extract_matrix(i, j);
+        return res;
+    } catch(std::exception &ex) {
+        forward_exception_to_r(ex);
+        return 0;
+    }
+};