+++ tests/testthat/test-BEDMatrix.R
@@ -1,25 +1,4 @@
-context("BEDMatrix")
-
-parseRaw <- function(path) {
-    file <- file(path)
-    lines <- readLines(file)
-    out <- t(sapply(strsplit(lines[2:length(lines)], " "), function(line) {
-        line <- line[7:length(line)]
-        line <- gsub("NA", NA, line)
-        as.integer(line)
-    }))
-    rownames(out) <- paste0("id_", 1:nrow(out))
-    colnames(out) <- paste0("mrk_", 1:ncol(out))
-    close(file)
-    return(out)
-}
-
-# Prepare dummy data
-raw <- parseRaw(system.file("extdata", "example.raw", package = "BEDMatrix"))
-examplePath <- system.file("extdata", "example.bed", package = "BEDMatrix")
-standalonePath <- "standalone.bed"
-
-for (path in c(examplePath, sub(".bed", "", examplePath))) {
+for (path in c(paste0(extdataPath, "/example"), paste0(extdataPath, "/example.bed"))) {
 
     test_that("it throws an error if file does not exist", {
         expect_error(BEDMatrix("NOT_FOUND"), "File not found\\.")
@@ -31,43 +10,43 @@ for (path in c(examplePath, sub(".bed", "", examplePath))) {
 
     test_that("it determines n from FAM file", {
         bed <- BEDMatrix(path = path)
-        expect_equal(nrow(bed), 6)
-        expect_message(BEDMatrix(path = path), "Extracting number of individuals and rownames from FAM file\\.\\.\\.")
+        expect_equal(nrow(bed), nrow(CROCHET_EXTRACT_ENV$COMPARE_OBJECT))
+        expect_message(BEDMatrix(path = path), "Extracting number of samples and rownames from FAM file\\.\\.\\.")
     })
 
     test_that("it throws an error if FAM file is not found and n is not given", {
-        expect_error(BEDMatrix(path = standalonePath), "FAM file of same name not found\\. Provide number of individuals \\(n\\)\\.")
+        expect_error(BEDMatrix(path = standalonePath), "FAM file of same name not found\\. Provide number of samples \\(n\\)\\.")
     })
 
     test_that("it determines rownames from FAM file", {
         bed <- BEDMatrix(path = path)
-        expect_equal(rownames(bed), c("1_1", "1_2", "1_3", "2_1", "2_2", "2_3"))
-        expect_message(BEDMatrix(path = path), "Extracting number of individuals and rownames from FAM file\\.\\.\\.")
+        expect_equal(rownames(bed), rownames(CROCHET_EXTRACT_ENV$COMPARE_OBJECT))
+        expect_message(BEDMatrix(path = path), "Extracting number of samples and rownames from FAM file\\.\\.\\.")
     })
 
     test_that("it determines p from BIM file", {
         bed <- BEDMatrix(path = path)
-        expect_equal(ncol(bed), 3)
-        expect_message(BEDMatrix(path = path), "Extracting number of markers and colnames from BIM file\\.\\.\\.")
+        expect_equal(ncol(bed), ncol(CROCHET_EXTRACT_ENV$COMPARE_OBJECT))
+        expect_message(BEDMatrix(path = path), "Extracting number of variants and colnames from BIM file\\.\\.\\.")
     })
 
     test_that("it throws an error if BIM file is not found and p is not given", {
-        expect_error(BEDMatrix(path = standalonePath), "FAM file of same name not found\\. Provide number of individuals \\(n\\)\\.")
+        expect_error(BEDMatrix(path = standalonePath), "FAM file of same name not found\\. Provide number of samples \\(n\\)\\.")
     })
 
     test_that("it determines colnames from BIM file", {
         bed <- BEDMatrix(path = path)
-        expect_equal(colnames(bed), c("snp1_G", "snp2_1", "snp3_A"))
-        expect_message(BEDMatrix(path = path), "Extracting number of markers and colnames from BIM file\\.\\.\\.")
+        expect_equal(colnames(bed), colnames(CROCHET_EXTRACT_ENV$COMPARE_OBJECT))
+        expect_message(BEDMatrix(path = path), "Extracting number of variants and colnames from BIM file\\.\\.\\.")
     })
 
     test_that("it accepts n and p if FAM or BIM file are present", {
-        bed <- BEDMatrix(path = path, n = 6, p = 3)
+        bed <- BEDMatrix(path = path, n = nrow(CROCHET_EXTRACT_ENV$COMPARE_OBJECT), p = ncol(CROCHET_EXTRACT_ENV$COMPARE_OBJECT))
         expect_equal(dimnames(bed), list(NULL, NULL))
     })
 
     test_that("it accepts n and p if FAM or BIM file is not found", {
-        bed <- BEDMatrix(path = standalonePath, n = 6, p = 3)
+        bed <- BEDMatrix(path = standalonePath, n = 3, p = 6)
         expect_equal(dimnames(bed), list(NULL, NULL))
     })
 
@@ -76,110 +55,3 @@ for (path in c(examplePath, sub(".bed", "", examplePath))) {
     })
 
 }
-
-# Prepare dummy BED matrix
-bed <- BEDMatrix(path = examplePath, n = 6, p = 3)
-rownames(bed) <- paste0("id_", 1:nrow(bed))
-colnames(bed) <- paste0("mrk_", 1:ncol(bed))
-
-test_that("subsetting", {
-
-    expect_equal(bed[], raw[])
-    expect_equal(typeof(bed[]), "integer")
-
-    expect_equal(bed[1, ], raw[1, ])
-    expect_equal(bed[, 1], raw[, 1])
-    expect_equal(bed[1, 1], raw[1, 1])
-    expect_equal(bed[1, , drop = FALSE], raw[1, , drop = FALSE])
-    expect_equal(bed[, 1, drop = FALSE], raw[, 1, drop = FALSE])
-    expect_equal(bed[1, 1, drop = FALSE], raw[1, 1, drop = FALSE])
-    expect_equal(typeof(bed[1, ]), "integer")
-
-    expect_equal(bed[1:2, ], raw[1:2, ])
-    expect_equal(bed[, 1:2], raw[, 1:2])
-    expect_equal(bed[1:2, 1:2], raw[1:2, 1:2])
-    expect_equal(bed[1:2, , drop = FALSE], raw[1:2, , drop = FALSE])
-    expect_equal(bed[, 1:2, drop = FALSE], raw[, 1:2, drop = FALSE])
-    expect_equal(bed[1:2, 1:2, drop = FALSE], raw[1:2, 1:2, drop = FALSE])
-    expect_equal(typeof(bed[1:2, ]), "integer")
-
-    expect_equal(bed[2:1, ], raw[2:1, ])
-    expect_equal(bed[, 2:1], raw[, 2:1])
-    expect_equal(bed[2:1, 2:1], raw[2:1, 2:1])
-    expect_equal(bed[2:1, , drop = FALSE], raw[2:1, , drop = FALSE])
-    expect_equal(bed[, 2:1, drop = FALSE], raw[, 2:1, drop = FALSE])
-    expect_equal(bed[2:1, 2:1, drop = FALSE], raw[2:1, 2:1, drop = FALSE])
-    expect_equal(typeof(bed[2:1, ]), "integer")
-
-    expect_equal(bed[c(3, 1), ], raw[c(3, 1), ])
-    expect_equal(bed[, c(3, 1)], raw[, c(3, 1)])
-    expect_equal(bed[c(3, 1), c(3, 1)], raw[c(3, 1), c(3, 1)])
-    expect_equal(bed[c(3, 1), , drop = FALSE], raw[c(3, 1), , drop = FALSE])
-    expect_equal(bed[, c(3, 1), drop = FALSE], raw[, c(3, 1), drop = FALSE])
-    expect_equal(bed[c(3, 1), c(3, 1), drop = FALSE], raw[c(3, 1), c(3, 1), drop = FALSE])
-    expect_equal(typeof(bed[c(3, 1), ]), "integer")
-
-    expect_equal(bed[c(TRUE, FALSE, TRUE), ], raw[c(TRUE, FALSE, TRUE), ])
-    expect_equal(bed[, c(TRUE, FALSE, TRUE)], raw[, c(TRUE, FALSE, TRUE)])
-    expect_equal(bed[c(TRUE, FALSE, TRUE), c(TRUE, FALSE, TRUE)], raw[c(TRUE, FALSE, TRUE), c(TRUE, FALSE, TRUE)])
-    expect_equal(bed[c(TRUE, FALSE, TRUE), , drop = FALSE], raw[c(TRUE, FALSE, TRUE), , drop = FALSE])
-    expect_equal(bed[, c(TRUE, FALSE, TRUE), drop = FALSE], raw[, c(TRUE, FALSE, TRUE), drop = FALSE])
-    expect_equal(bed[c(TRUE, FALSE, TRUE), c(TRUE, FALSE, TRUE), drop = FALSE], raw[c(TRUE, FALSE, TRUE), c(TRUE, FALSE, TRUE), drop = FALSE])
-    expect_equal(typeof(bed[c(TRUE, FALSE, TRUE), ]), "integer")
-
-    expect_equal(bed[1], raw[1])
-    expect_equal(bed[1:2], raw[1:2])
-    expect_equal(bed[2:1], raw[2:1])
-    expect_equal(bed[c(3, 1)], raw[c(3, 1)])
-    expect_equal(bed[c(TRUE, FALSE, TRUE)], raw[c(TRUE, FALSE, TRUE)])
-    expect_equal(bed[raw > 1], raw[raw > 1])
-    expect_equal(typeof(bed[1]), "integer")
-
-    expect_equal(bed["id_1", ], raw["id_1", ])
-    expect_equal(bed[, "mrk_1"], raw[, "mrk_1"])
-    expect_equal(bed["id_1", "mrk_1"], raw["id_1", "mrk_1"])
-    expect_equal(bed["id_1", , drop = FALSE], raw["id_1", , drop = FALSE])
-    expect_equal(bed[, "mrk_1", drop = FALSE], raw[, "mrk_1", drop = FALSE])
-    expect_equal(bed["id_1", "mrk_1", drop = FALSE], raw["id_1", "mrk_1", drop = FALSE])
-    expect_equal(typeof(bed["id_1", ]), "integer")
-
-    expect_equal(bed[c("id_1", "id_2"), ], raw[c("id_1", "id_2"), ])
-    expect_equal(bed[, c("mrk_1", "mrk_2")], raw[, c("mrk_1", "mrk_2")])
-    expect_equal(bed[c("id_1", "id_2"), c("mrk_1", "mrk_2")], raw[c("id_1", "id_2"), c("mrk_1", "mrk_2")])
-    expect_equal(bed[c("id_1", "id_2"), , drop = FALSE], raw[c("id_1", "id_2"), , drop = FALSE])
-    expect_equal(bed[, c("mrk_1", "mrk_2"), drop = FALSE], raw[, c("mrk_1", "mrk_2"), drop = FALSE])
-    expect_equal(bed[c("id_1", "id_2"), c("mrk_1", "mrk_2"), drop = FALSE], raw[c("id_1", "id_2"), c("mrk_1", "mrk_2"), drop = FALSE])
-    expect_equal(typeof(bed[c("id_1", "id_2"), ]), "integer")
-
-    expect_equal(bed[c("id_2", "id_1"), ], raw[c("id_2", "id_1"), ])
-    expect_equal(bed[, c("mrk_2", "mrk_1")], raw[, c("mrk_2", "mrk_1")])
-    expect_equal(bed[c("id_2", "id_1"), c("mrk_2", "mrk_1")], raw[c("id_2", "id_1"), c("mrk_2", "mrk_1")])
-    expect_equal(bed[c("id_2", "id_1"), , drop = FALSE], raw[c("id_2", "id_1"), , drop = FALSE])
-    expect_equal(bed[, c("mrk_2", "mrk_1"), drop = FALSE], raw[, c("mrk_2", "mrk_1"), drop = FALSE])
-    expect_equal(bed[c("id_2", "id_1"), c("mrk_2", "mrk_1"), drop = FALSE], raw[c("id_2", "id_1"), c("mrk_2", "mrk_1"), drop = FALSE])
-    expect_equal(typeof(bed[c("id_2", "id_1"), ]), "integer")
-
-    expect_equal(bed[c("id_3", "id_1"), ], raw[c("id_3", "id_1"), ])
-    expect_equal(bed[, c("mrk_3", "mrk_1")], raw[, c("mrk_3", "mrk_1")])
-    expect_equal(bed[c("id_3", "id_1"), c("mrk_3", "mrk_1")], raw[c("id_3", "id_1"), c("mrk_3", "mrk_1")])
-    expect_equal(bed[c("id_3", "id_1"), , drop = FALSE], raw[c("id_3", "id_1"), , drop = FALSE])
-    expect_equal(bed[, c("mrk_3", "mrk_1"), drop = FALSE], raw[, c("mrk_3", "mrk_1"), drop = FALSE])
-    expect_equal(bed[c("id_3", "id_1"), c("mrk_3", "mrk_1"), drop = FALSE], raw[c("id_3", "id_1"), c("mrk_3", "mrk_1"), drop = FALSE])
-    expect_equal(typeof(bed[c("id_3", "id_1"), ]), "integer")
-
-    # Do not modify indexes
-    i <- seq_len(6 * 3)
-    bed[i]
-    expect_equal(i, seq_len(6 * 3))
-
-    i <- seq_len(6)
-    j <- seq_len(3)
-    bed[i, j]
-    expect_equal(i, seq_len(6))
-    expect_equal(j, seq_len(3))
-
-})
-
-test_that("length", {
-    expect_equal(length(bed), length(raw))
-})