+++ tests/testthat/test-length.R
@@ -1,0 +1,3 @@
+test_that("length", {
+    expect_equal(length(CROCHET_EXTRACT_ENV$CUSTOM_OBJECT), length(CROCHET_EXTRACT_ENV$COMPARE_OBJECT))
+})